requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=10.0.0
matplotlib>=3.4.0
scikit-image>=0.19.0
tqdm>=4.65.0
cryptography>=3.4.0
pyOpenSSL>=23.0.0
scapy>=2.6.0
tk
reportlab==4.0.4
markdown==3.4.3
beautifulsoup4==4.12.2
